Melting
Melting snow is now working properly.
Found a way to get the absolute coordinates of a block by using  'chunk.getPos().getStartPos()'.

This allowed me to use 'world.setBlockState' instead of 'chunk.setBlockState'.

It also allowed me to use 'onBlockUpdate' for a client, this fixed the client not rendering the new block state.

-Moved 'meltBlocks' from 'OpenSeasonsMod' to 'WorldManipulation'

-Moved 'updateClientBlock' from OpenSeasonsMod to 'S2C'