Reworked water and ice melting.
Using mixins to inject correct precipitation, snow and ice generation, snow and ice melting, according to the current season

Key values for loading and saving from the openseasons.json file moved to 'openseasons.util.Keys' class

Removed CLIENT_JOIN global receiver

updated temperatures to match existing vanilla minecraft biomes.

Unfortunately this required the use of a variable to store the current Season so the biomes can know how to behave properly, this means worlds are not independent as far as precipitation and block melting are concerned. Because of that, these features will depend on 'enable_dynamic_weather' being set to true in the 'openseasons.json' file.