Fixed precipitation
Precipitation is now rendering properly, according to the current season instead of the biome temperature.

Decided to scrap most of the injections in 'BiomeMixin' to avoid messing with world generation.

Since all I wanted was for water, ice and snow to generate dynamically, all of these tasks are now being injected directly in 'ServerWorld.tickChunk'